Handled more edge cases for the `superping` command.
